+++ lib/modern_player.dart
@@ -1,11 +1,17 @@
 library modern_player;
 
 export 'src/modern_player.dart' show ModernPlayer;
-export 'src/modern_players_enums.dart' show ModernPlayerType;
+export 'src/modern_players_enums.dart'
+    show
+        ModernPlayerSourceType,
+        ModernPlayerSubtitleSourceType,
+        ModernPlayerAudioSourceType;
 export 'src/modern_player_options.dart'
     show
         ModernPlayerQualityOptions,
         ModernPlayerControlsOptions,
         ModernPlayerThemeOptions,
         ModernPlayerProgressSliderTheme,
-        ModernPlayerCustomActionButton;
+        ModernPlayerCustomActionButton,
+        ModernPlayerSubtitleOptions,
+        ModernPlayerAudioTrackOptions;